+++ integration_tests/suite/test_presences.py
@@ -23,9 +23,9 @@ from wazo_chatd_client.exceptions import ChatdError
 from .helpers import fixtures
 from .helpers.base import (
     BaseIntegrationTest,
+    TOKEN_SUBTENANT_UUID,
+    TOKEN_TENANT_UUID,
     UNKNOWN_UUID,
-    MASTER_TENANT_UUID,
-    SUBTENANT_UUID,
 )
 
 USER_UUID = str(uuid.uuid4())
@@ -66,8 +66,8 @@ class TestPresence(BaseIntegrationTest):
             filtered=equal_to(2),
         ))
 
-    @fixtures.db.user(tenant_uuid=MASTER_TENANT_UUID)
-    @fixtures.db.user(tenant_uuid=SUBTENANT_UUID)
+    @fixtures.db.user(tenant_uuid=TOKEN_TENANT_UUID)
+    @fixtures.db.user(tenant_uuid=TOKEN_SUBTENANT_UUID)
     def test_list_multi_tenant(self, user_1, user_2):
         presences = self.chatd.user_presences.list()
         assert_that(presences, has_entries(
@@ -141,15 +141,15 @@ class TestPresence(BaseIntegrationTest):
             )
         )
 
-    @fixtures.db.user(tenant_uuid=MASTER_TENANT_UUID)
-    @fixtures.db.user(tenant_uuid=SUBTENANT_UUID)
+    @fixtures.db.user(tenant_uuid=TOKEN_TENANT_UUID)
+    @fixtures.db.user(tenant_uuid=TOKEN_SUBTENANT_UUID)
     def test_get_multi_tenant(self, user_1, user_2):
-        result = self.chatd.user_presences.get(user_2.uuid, tenant_uuid=SUBTENANT_UUID)
+        result = self.chatd.user_presences.get(user_2.uuid, tenant_uuid=TOKEN_SUBTENANT_UUID)
         assert_that(result, has_entries(uuid=user_2.uuid))
 
         assert_that(
             calling(self.chatd.user_presences.get).with_args(
-                user_1.uuid, tenant_uuid=SUBTENANT_UUID,
+                user_1.uuid, tenant_uuid=TOKEN_SUBTENANT_UUID,
             ),
             raises(ChatdError, has_properties(status_code=404))
         )
@@ -184,18 +184,18 @@ class TestPresence(BaseIntegrationTest):
             )
         )
 
-    @fixtures.db.user(tenant_uuid=MASTER_TENANT_UUID)
-    @fixtures.db.user(tenant_uuid=SUBTENANT_UUID, state='unavailable')
+    @fixtures.db.user(tenant_uuid=TOKEN_TENANT_UUID)
+    @fixtures.db.user(tenant_uuid=TOKEN_SUBTENANT_UUID, state='unavailable')
     def test_update_multi_tenant(self, user_1, user_2):
         user_args = {'uuid': user_2.uuid, 'state': 'available'}
-        self.chatd.user_presences.update(user_args, tenant_uuid=SUBTENANT_UUID)
+        self.chatd.user_presences.update(user_args, tenant_uuid=TOKEN_SUBTENANT_UUID)
 
         result = self.chatd.user_presences.get(user_args['uuid'])
         assert_that(result, has_entries(user_args))
 
         assert_that(
             calling(self.chatd.user_presences.update).with_args(
-                {'uuid': user_1.uuid}, tenant_uuid=SUBTENANT_UUID,
+                {'uuid': user_1.uuid}, tenant_uuid=TOKEN_SUBTENANT_UUID,
             ),
             raises(ChatdError, has_properties(status_code=404))
         )